==> Data_analytics/1. merge instances of route.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.patches as mpatches
import matplotlib.font_manager as fm
import matplotlib
import pickle
import timeit
import h5py
import math
import os
import json
# import seaborn as sns
%matplotlib inline
import datetime
from os import listdir
from os.path import isfile, join

path = '/Users/jakek/Documents/Computer Science/Semester 3/Data Analytics/compiled compiled 2/'
import os

def clean(file):
    features = list(file.columns.values)
    if 'Unnamed: 0' in features:
        file.drop('Unnamed: 0', axis=1, inplace=True)
    if 'Unnamed: 0.1' in features:
        file.drop('Unnamed: 0.1', axis=1, inplace=True) 
    
    if 'null' in file['StopID'].unique():
        file = file[file.StopID != 'null']
    file['StopID'] = file['StopID'].astype(float).astype(int)
    
    file = file.groupby(['VehicleJourneyID','TimeFrame']).filter(lambda x: (x.StopID).nunique() > 8)
    value = frequent(file)
    file = file.groupby(['VehicleJourneyID','TimeFrame']).filter(lambda x: x['StopID'].loc[x['Timestamp'] == x['Timestamp'].min()].values[0] == value)
    return file

def frequent(file):
    starts = []
    for name, group in file.groupby(['VehicleJourneyID','TimeFrame']):
            starts+=[group[group.Timestamp == group.Timestamp.min()].StopID.values[0]]
    if len(starts)>0:
        return max(set(starts), key=starts.count)
    else:
        return None

files = []
instances=[]
for file in  os.listdir(path):
    instances += [file[4:-4]]
    file = file[3:8]
    if file[0] =='0':
        file = file[1:]
        files += [file]
files=set(files)

print(files)

for file in files:
    compiled = []
    for i in instances:
        if file == i[:4]:
            df = pd.read_csv(path+ 'bus0'+i+'.csv', low_memory= False)
            df.drop('Unnamed: 0', axis=1, inplace=True)
            df.drop('Unnamed: 0.1', axis=1, inplace=True)
            df.drop('Unnamed: 0.1.1', axis=1, inplace=True)
            df = clean(df)
            df['Instance'] = int(i[-1])
            compiled +=[df]
    complete = pd.concat(compiled)
    complete['Direction'] = file[-1]
    complete.to_csv("./complete dfs/"+file+".csv")
    print(file, complete.shape[0])

### old code

df.drop('Unnamed: 0', axis=1, inplace=True)
df.drop('Unnamed: 0.1', axis=1, inplace=True)
df.drop('Unnamed: 0.1.1', axis=1, inplace=True)
df = clean(df)

df1.drop('Unnamed: 0', axis=1, inplace=True)
df1.drop('Unnamed: 0.1', axis=1, inplace=True)
df1.drop('Unnamed: 0.1.1', axis=1, inplace=True)
df1 = clean(df1)

df['Instance'] = 1
df1['Instance'] = 2


complete = pd.concat([df,df1])

complete['Direction'] = 1

complete.head()

complete.to_csv("./complete dfs/bus 4/4_1.csv")

complete.Instance.unique()
