==> Data_analytics/3. Merge directions (CSV and JSON).ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.patches as mpatches
import matplotlib.font_manager as fm
import matplotlib
import pickle
import timeit
import h5py
import math
import os
import json
# import seaborn as sns
%matplotlib inline
import datetime
from os import listdir
import json
from os.path import isfile, join

files = []
routes = []
for file in  os.listdir("./complete dfs/cleaned/"):
    if isfile("./complete dfs/cleaned/"+file):
        routes += [file[:3]]
        files += [file[:4]]
files=set(files)
routes = set(routes)

print(len(files),  len(routes))

ready = []
alone = []
for route in routes:
    i=0
    for file in files:
        if route in file[:-1]:
            i+=1
    if i >1:
        ready +=[route]
    elif i < 2:
        alone += [route]
print(len(ready))

print(alone)

print(files)

# for csv

for i in ready:
    print(i)
    if isfile("./complete dfs/cleaned and merged/"+i+'.csv'):
        print("already exists")
    else:
        dfs = []
        for file in files:
            if i in file[:-1]:
                df = pd.read_csv('./complete dfs/cleaned/'+ file+'cleaned.csv', low_memory= False)
                dfs+= [df]
        size = []
        for j in dfs:
            size += [j.shape[1]]
        if len(set(size)) != 1:
            print("ERROR")
        else:
            complete = pd.concat(dfs)
            complete.to_csv("./complete dfs/cleaned and merged/"+i+".csv")
                
        
        

# same for json
dicty = {}
for i in ready:
    print(i)
    if isfile("./JSON2/merged/"+i+'.json'):
        print("already exists")
    else:
        dicty[i] = {}
        for file in files:
            if i in file[:-1]:
                  with open('./JSON2/'+file+'.json') as data_file:
                        dicty[i][file[-1]] = json.load(data_file)
        with open('./JSON2/merged/'+i+'.json', 'w') as fp:
            json.dump(dicty[i], fp)



# now for the ones that dont need to be merged

# for csv

for i in alone:
    print(i)
    if isfile("./complete dfs/cleaned and merged/"+i+'.csv'):
        print("already exists")
    else:
        dfs = []
        for file in files:
            if i in file[:-1]:
                df = pd.read_csv('./complete dfs/cleaned/'+ file+'cleaned.csv', low_memory= False)
                dfs+= [df]
        size = []
        for j in dfs:
            size += [j.shape[1]]
        if len(set(size)) != 1:
            print("ERROR")
        else:
            complete = pd.concat(dfs)
            complete.to_csv("./complete dfs/cleaned and merged/"+i+".csv")
                
        
        

# same for json
dicty = {}
for i in alone:
    print(i)
    if isfile("./JSON2/merged/"+i+'.json'):
        print("already exists")
    else:
        dicty[i] = {}
        for file in files:
            if i in file[:-1]:
                  with open('./JSON2/'+file+'.json') as data_file:
                        dicty[i][file[-1]] = json.load(data_file)
        with open('./JSON2/merged/'+i+'.json', 'w') as fp:
            json.dump(dicty[i], fp)



## old code

import math
math.isnan(dicty['31B']['1']['5']['10']['621'])

# cant figure this one out

remove = []
for bus in dicty:
    for direction in dicty[bus]:
        for day in dicty[bus][direction]:
            if dicty[bus][direction][day] == 'index':
                continue
            else:
                for time in dicty[bus][direction][day]:
                    if dicty[bus][direction][day][time] == 'index':
                        continue
                    else:
                        print(time)
                        for stop in dicty[bus][direction][day][time]:
                            if math.isnan(dicty[bus][direction][day][time][stop]):
                                remove +=[bus, direction, day, time, stop]


path = '/Users/jakek/Documents/Computer Science/Semester 3/Data Analytics/Working folder/complete dfs/bus 1/'
df = pd.read_csv(path+ '4_1cleaned.csv', low_memory= False)
df1 = pd.read_csv(path+ '4_0cleaned.csv', low_memory= False)


df.shape

df1.shape

complete = pd.concat([df,df1])

complete.shape

complete.to_csv("./complete dfs/cleaned and merged/bus 4 complete.csv")

with open('./JSON2/bus1/4_0.json') as data_file:    
    dir_0 = json.load(data_file)
    
with open('./JSON2/bus1/4_1.json') as data_file:    
    dir_1 = json.load(data_file)

dir_0['3']['12']

dir_1['3']['12']

bus_4 = {}

bus_4['0']= dir_0
bus_4['1']= dir_1

with open('./JSON2/bus1/Bus_4.json', 'w') as fp:
    json.dump(bus_4, fp)
