==> Data_analytics/3.5 merge multiple routes (add route to CSV and merge JSON).ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.patches as mpatches
import matplotlib.font_manager as fm
import matplotlib
import pickle
import timeit
import math
import os
import json
# import seaborn as sns
%matplotlib inline
import datetime
from os import listdir
from os.path import isfile, join
import winsound

df4 = pd.read_csv('./complete dfs/bus 1/bus 4 complete.csv', low_memory= False)
df1 = pd.read_csv('./complete dfs/bus 1/bus 1 complete.csv', low_memory= False)

df1['Route'] = 1
df4['Route'] = 4

df = pd.concat([df1,df4])

dfs = []
for file in  os.listdir("./complete dfs/cleaned and merged/"):
    df = pd.read_csv('./complete dfs/cleaned and merged/'+file, low_memory= False)
    df = df.dropna()
    df['Route'] = file[:3]
    dfs += [df]
all = pd.concat(dfs)


dfs[0].head()

all.to_csv("./complete dfs/complete compile/hunormus.csv")

Freq = 2500 
Dur = 1000 
winsound.Beep(Freq,Dur)

Freq = 2500 
Dur = 1000 
winsound.Beep(Freq,Dur)

all.shape

all.head()

twenty = all.sample(frac=0.2)

twenty.head()

twenty.to_csv("./complete dfs/complete compile/twenty.csv")
