==> Data_analytics/4. modeling.ipynb <==
import numpy as np
import pandas as pd
# import random as rnd
# import seaborn as sns
# import matplotlib.pyplot as plt
# from patsy import dmatrices
from sklearn.cross_validation import train_test_split
from sklearn import metrics
from sklearn.cross_validation import cross_val_score
# from sklearn.tree import export_graphviz
# from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import RandomForestRegressor

# import winsound
%matplotlib inline
import datetime

file = 'twenty'

df_classif = pd.read_csv('complete dfs/Complete compile/'+file+'.csv',low_memory= False)

# df2.replace('Monday',1,inplace=True)
# df2.replace('Tuesday',2,inplace=True)
# df2.replace('Wednesday',3,inplace=True)
# df2.replace('Thursday',4,inplace=True)
# df2.replace('Friday',5,inplace=True)
# df2.replace('Saturday',6,inplace=True)
# df2.replace('Sunday',7,inplace=True)

# df_classif = pd.concat([df_classif,df2])


df_classif.head()

df_classif.replace('Clear',1,inplace=True)
df_classif.replace('Fog',2,inplace=True)
df_classif.replace('Light Drizzle',3,inplace=True)
df_classif.replace('Light Rain',4,inplace=True)
df_classif.replace('Light Rain Showers',5,inplace=True)
df_classif.replace('Light Small Hail Showers',6,inplace=True)
df_classif.replace('Light Snow Showers',7,inplace=True)
df_classif.replace('Mostly Cloudy',8,inplace=True)
df_classif.replace('Overcast',9,inplace=True)
df_classif.replace('Partly Cloudy',10,inplace=True)
df_classif.replace('Patches of Fog',11,inplace=True)
df_classif.replace('Rain',12,inplace=True)
df_classif.replace('Scattered Clouds',13,inplace=True)
df_classif.replace('Shallow Fog',14,inplace=True)

df_classif.Direction.unique()

df_classif=df_classif[df_classif.Avg > 0]
df_classif=df_classif[df_classif.TimeToNextStop> 0]

def mins(x):
    x = x/1000000
    date = datetime.datetime.fromtimestamp(x).strftime("%H, %M")
    hour, minute = date.split(', ')
    date = hour+'.'+minute
    return(float(date))


df_classif['mins'] = df_classif.apply(lambda row: mins(row['Timestamp']), axis=1)

df_classif[['VehicleJourneyID','StopID','Timestamp','AtStop','Avg','TimeToNextStop','Day','Summary','Temp','Rain','Wind_Speed', 'Direction', 'mins']] = df_classif[['VehicleJourneyID','StopID','Timestamp','AtStop','Avg','TimeToNextStop','Day','Summary','Temp','Rain','Wind_Speed', 'Direction', 'mins']].astype(np.float32)
df_classif=df_classif.dropna()
df_classif.head()

# num = []
# for i in df_classif['Route'].unique():
#     try:
#         int(i)
#         num += [i]
#     except:
#         print(i, "is not int")

# testroute = df_classif.loc[df_classif['Route'] == num[0]]
# for i in num:
#     if i != num[0]:
#         testroute = pd.concat([testroute, df_classif.loc[df_classif['Route'] == i]]) 

df_cont_feat = df_classif[['Avg','Temp']]
df_dummies_broadband = df_classif[['StopID','AtStop','Day', 'mins']]
df_dummies_broadband.head()

X = pd.concat([df_cont_feat, df_dummies_broadband[['StopID','AtStop','Day', 'mins']]], axis =1)
y = df_classif.TimeToNextStop

from sklearn import preprocessing
from sklearn import utils

lab_enc = preprocessing.LabelEncoder()
y = lab_enc.fit_transform(y)

rfr = RandomForestRegressor(n_estimators=15, max_features='auto', max_depth=50, oob_score=True, random_state=1, min_samples_split=100)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3)


rfr.fit(X_train, y_train)


rfr_predictions_test = rfr.predict(X_test)
df_true_vs_rfr_predicted_test = pd.DataFrame({'Actual': y_test, 'Predicte': rfr_predictions_test})
df_true_vs_rfr_predicted_test.head()

print("Accuracy: ", metrics.accuracy_score(y_test, rfc_predictions_test))


new4=abs(df_true_vs_rfr_predicted_test.Actual-df_true_vs_rfr_predicted_test.Predicte)

total=0
for i in new4:
    if i<=20:
        total+=1
accuracy=(total/len(new4))
print(accuracy)

pd.DataFrame({'feature': X_test.columns, 'importance':rfr.feature_importances_})

from sklearn.externals import joblib
joblib.dump(rfr, 'pickled rf/'+file+'extra time feature', compress=False)

import sys
sys.getsizeof(rfr)

X_test.head()

import winsound
while True:
    winsound.Beep(17000, 1000)

# df_classif = df_classif.sample(frac=0.4)

train, test = train_test_split(df_classif, test_size=0.2)
train = train.sample(frac=0.4)

train_X = train[['Avg','Temp', 'StopID','AtStop','Day','Hour']]
train_Y = train.TimeToNextStop
test_X = test[['Avg','Temp', 'StopID','AtStop','Day','Hour']]
test_Y = test.TimeToNextStop

lab_enc = preprocessing.LabelEncoder()
train_Y = lab_enc.fit_transform(train_Y)
test_Y = lab_enc.fit_transform(test_Y)

rfc.fit(train_X, train_Y)

rfc_predictions_test = rfc.predict(test_X)

df_true_vs_rfc_predicted_test = pd.DataFrame({'Actual': test_Y, 'Predicte': rfc_predictions_test})
df_true_vs_rfc_predicted_test.head()

print("Accuracy: ", metrics.accuracy_score(test_Y, rfc_predictions_test))
new4=abs(df_true_vs_rfc_predicted_test.Actual-df_true_vs_rfc_predicted_test.Predicte)

total=0
for i in new4:
    if i<=20:
        total+=1
accuracy=(total/len(new4))
print(accuracy)

joblib.dump(rfc, 'pickled rf/'+file+' 40percentoftrain.pkl', compress=True)
